20 june

+++ ClubJumana.Core/DTOs/ProductInformationViewModel.cs
@@ -79,10 +79,10 @@ namespace ClubJumana.Core.DTOs
         public int? ColourFK { get; set; }
         public int? BarcodeFK { get; set; }
         public int? ProductTypeFK { get; set; }
-        public float? FobPrice { get; set; }
-        public float? WholesaleA { get; set; }
-        public float? WholesaleB { get; set; }
-        public float? RetailPrice { get; set; }
+        public decimal? FobPrice { get; set; }
+        public decimal? WholesaleA { get; set; }
+        public decimal? WholesaleB { get; set; }
+        public decimal? RetailPrice { get; set; }
         public string Size { get; set; }
         public Product Product { get; set; }
         public Colour Colour { get; set; }
